Restructured files and added new folders
