+++ index.html
@@ -13,8 +13,7 @@
       {
         "imports": {
           "three": "../../lib/three.js-r145/build/three.module.js",
-          "controls": "../../lib/three.js-r145/examples/jsm/controls/OrbitControls.js",
-          "datgui": "../../lib/dat.gui-0.7.9/build/dat.gui.module.js"
+          "controls": "../../lib/three.js-r145/examples/jsm/controls/OrbitControls.js"
         }
       }
     </script>
